DsearcheR
DsearcheR is the name of a tool for all that needs be done by any computer.  Searching and finding are the main acts of anything there is for moving.  Some may bend such a reality for a so called better this and that; somewhere somehow at this and those movement, it finds itself doing the same thing it denied itself doing. (Mt7,7;Mt11,29;Lk11,9)